prevent sign expansion problems with 64bit platforms that define the ioctl request parameter as long (*BSD)

+++ libmetrics/get_ifi_info.c
@@ -36,7 +36,7 @@ Socket(int family, int type, int protocol)
 }
 
 int
-Ioctl(int fd, int request, void *arg)
+Ioctl(int fd, unsigned long request, void *arg)
 {
   int             n;
